modify bugs

+++ src/abs/services/crm/university/manager/__init__.py
@@ -183,7 +183,7 @@ class UniversityServer(BaseManager):
         return major
 
     @classmethod
-    def remove_major(cls,):
+    def remove_major(cls, major_id):
         major = cls.get_major(major_id)
         relations_qs = UniversityRelationsServer.search_all(major = major)
         if relations_qs.count() > 0:
@@ -239,6 +239,11 @@ class UniversityRelationsServer(BaseManager):
     @classmethod
     def remove(cls, relations_id):
         relations = cls.get(relations_id)
+        years_qs = UniversityYearsServer.search_all(
+            relations = relations
+        )
+        if years_qs.count() > 0:
+            raise BusinessError("存在学年无法删除")
         relations.delete()
         return True
